requirements.txt
langchain
langgraph
faiss-cpu
sentence-transformers
requests
python-dotenv